Move vehicleFormSchema to co-located _schemas.ts
The schema was previously defined in `vehicles/new/+page.server.ts` and
imported via a relative path from the edit page. Extracting it to
`web-v4/src/routes/vehicles/_schemas.ts` (SvelteKit ignores the `_`
prefix for routing) keeps the schema co-located with the pages that
use it, while removing the awkward `../../new/+page.server` import.
